Fixes issue with celery bootup by using the right signal, adds test.
