+++ tests/test_seer.py
@@ -8,7 +8,6 @@ from johen import generate
 from johen.pytest import parametrize
 from sqlalchemy import text
 
-from celery_app.app import celery_app
 from seer.app import app
 from seer.automation.autofix.models import AutofixContinuation, AutofixEvaluationRequest
 from seer.automation.codebase.models import CodebaseStatusCheckRequest, CodebaseStatusCheckResponse
@@ -633,16 +632,3 @@ class TestGetAutofixState:
         assert response.status_code == 200
         data = json.loads(response.get_data(as_text=True))
         assert data == {"group_id": None, "run_id": None, "state": None}
-
-
-def test_detected_celery_jobs():
-    assert set(k for k in celery_app.tasks.keys() if not k.startswith("celery.")) == set(
-        [
-            "seer.automation.autofix.steps.change_describer_step.autofix_change_describer_task",
-            "seer.automation.autofix.steps.coding_step.autofix_coding_task",
-            "seer.automation.autofix.steps.root_cause_step.root_cause_task",
-            "seer.automation.autofix.steps.steps.autofix_parallelized_chain_step_task",
-            "seer.automation.autofix.steps.steps.autofix_parallelized_conditional_step_task",
-            "seer.automation.autofix.tasks.run_autofix_evaluation_on_item",
-        ]
-    )
